+++ app/Console/Commands/CreateSuperuserCommand.php
@@ -2,12 +2,16 @@
 
 namespace App\Console\Commands;
 
-use App\Http\Requests\UserRequest;
 use App\Models\Role;
 use App\Models\User;
+use App\Rules\Password;
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Hash;
-use Illuminate\Support\Facades\Validator;
+use Illuminate\Validation\Rule;
+
+use function Laravel\Prompts\password;
+use function Laravel\Prompts\select;
+use function Laravel\Prompts\text;
 
 /**
  * Command class that makes it possible to create an new superuser.
@@ -65,36 +69,19 @@ class CreateSuperuserCommand extends Command
         $this->info('Creating an new superuser, please notify your inputs.');
 
         $data = [];
-        $data['firstname'] = $this->ask('Firstname');
-        $data['lastname'] = $this->ask('Lastname');
-        $data['email'] = $this->ask('E-Mail');
-        $data['user_locale'] = $this->ask('Locale (possible values: '.implode(',', array_keys(config('app.enabled_locales'))).')');
-        $data['password'] = $this->secret('Password');
-        $data['password_confirmation'] = $this->secret('Password Confirmation');
-        $data['generate_password'] = false;
-        $data['bbb_skip_check_audio'] = false;
-        $data['roles'] = [$superuserRole->id];
-        $data['timezone'] = 'UTC';
-
-        $validator = Validator::make($data, (new UserRequest)->rules());
-
-        if ($validator->fails()) {
-            $this->info('Something went wrong, please see the error messages below for more information.');
-
-            foreach ($validator->errors()->all() as $error) {
-                $this->error($error);
-            }
-
-            return 1;
-        }
+        $firstname = text('Firstname', validate: ['firstname' => 'required|max:255']);
+        $lastname = text('Lastname', validate: ['lastname' => 'required|max:255']);
+        $email = text('E-Mail', validate: ['email' => ['required', 'max:255', 'email', Rule::unique('users', 'email')->where('authenticator', 'local')]]);
+        $locale = select('Locale', array_keys(config('app.enabled_locales')));
+        $password = password('Password', required: true, validate: ['min:8', new Password]);
 
         $user = new User;
 
-        $user->firstname = $data['firstname'];
-        $user->lastname = $data['lastname'];
-        $user->email = $data['email'];
-        $user->locale = $data['user_locale'];
-        $user->password = Hash::make($data['password']);
+        $user->firstname = $firstname;
+        $user->lastname = $lastname;
+        $user->email = $email;
+        $user->locale = $locale;
+        $user->password = Hash::make($password);
         $user->email_verified_at = $user->freshTimestamp();
 
         $user->save();