+++ app/Http/Controllers/api/v1/RoleController.php
@@ -72,7 +72,10 @@ class RoleController extends Controller
         $role->superuser = false;
 
         $role->save();
-        $role->permissions()->sync($request->permissions);
+
+        $new_permissions = $role->filterRestrictedPermissions(collect($request->permissions));
+
+        $role->permissions()->sync($new_permissions);
 
         return (new RoleResource($role))->withPermissions();
     }
@@ -98,7 +101,9 @@ class RoleController extends Controller
         if (! $role->superuser) {
             $old_role_permissions = $role->permissions()->pluck('permissions.id')->toArray();
 
-            $role->permissions()->sync($request->permissions);
+            $new_permissions = $role->filterRestrictedPermissions(collect($request->permissions));
+
+            $role->permissions()->sync($new_permissions);
 
             $user = Auth::user();