+++ config/permissions.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'restrictions' => explode(',', env('PERMISSION_RESTRICTIONS', '')),
+];